+++ CHANGELOG.md
@@ -7,7 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## [Unreleased]
 
 ### Added
-- N/A
+- Ability to configure a post init message for geb to display once the template has been used. e.g. Used to remind users to update their geb.config.yml file
 
 ### Changed
 - N/A